Split test_integration_job (#92)
* split job integration tests

* fix lint
