+++ test/test_integration_program.py
@@ -12,13 +12,9 @@
 
 """Tests for runtime service."""
 
-import copy
 import unittest
 import os
-import uuid
-from contextlib import suppress
 import tempfile
-from collections import defaultdict
 
 from qiskit_ibm_runtime.exceptions import IBMNotAuthorizedError
 from qiskit_ibm_runtime.runtime_program import RuntimeProgram
@@ -26,35 +22,13 @@ from qiskit_ibm_runtime.exceptions import (
     RuntimeProgramNotFound,
 )
 
-from .ibm_test_case import IBMTestCase
-from .utils.decorators import requires_cloud_legacy_services, run_cloud_legacy_real
-from .utils.templates import RUNTIME_PROGRAM, RUNTIME_PROGRAM_METADATA, PROGRAM_PREFIX
+from .ibm_test_case import IBMIntegrationTestCase
+from .utils.decorators import run_cloud_legacy_real
+from .utils.templates import RUNTIME_PROGRAM, PROGRAM_PREFIX
 
 
-class TestIntegrationProgram(IBMTestCase):
+class TestIntegrationProgram(IBMIntegrationTestCase):
     """Integration tests for runtime modules."""
-
-    @classmethod
-    @requires_cloud_legacy_services
-    def setUpClass(cls, services):
-        """Initial class level setup."""
-        # pylint: disable=arguments-differ
-        super().setUpClass()
-        cls.services = services
-
-    def setUp(self) -> None:
-        """Test level setup."""
-        super().setUp()
-        self.to_delete = defaultdict(list)
-
-    def tearDown(self) -> None:
-        """Test level teardown."""
-        super().tearDown()
-        # Delete programs
-        for service in self.services:
-            for prog in self.to_delete[service.auth]:
-                with suppress(Exception):
-                    service.delete_program(prog)
 
     @run_cloud_legacy_real
     def test_list_programs(self, service):
@@ -210,7 +184,7 @@ def main(backend, user_messenger, **kwargs):
         program_id = self._upload_program(service)
         original = service.program(program_id)
         new_metadata = {
-            "name": self._get_program_name(),
+            "name": PROGRAM_PREFIX,
             "description": "test_update_program_metadata",
             "max_execution_time": original.max_execution_time + 100,
             "spec": {
@@ -233,27 +207,3 @@ def main(backend, user_messenger, **kwargs):
         self.assertTrue(program.max_execution_time)
         self.assertTrue(program.creation_date)
         self.assertTrue(program.update_date)
-
-    def _upload_program(
-        self,
-        service,
-        name=None,
-        max_execution_time=300,
-        data=None,
-        is_public: bool = False,
-    ):
-        """Upload a new program."""
-        name = name or self._get_program_name()
-        data = data or RUNTIME_PROGRAM
-        metadata = copy.deepcopy(RUNTIME_PROGRAM_METADATA)
-        metadata["name"] = name
-        metadata["max_execution_time"] = max_execution_time
-        metadata["is_public"] = is_public
-        program_id = service.upload_program(data=data, metadata=metadata)
-        self.to_delete[service.auth].append(program_id)
-        return program_id
-
-    @classmethod
-    def _get_program_name(cls):
-        """Return a unique program name."""
-        return PROGRAM_PREFIX + "_" + uuid.uuid4().hex
